começando a fazer tela de perfil

+++ app/src/main/java/br/com/pedrovieira/doetempo/models/Count.kt
@@ -1,6 +1,6 @@
 package br.com.pedrovieira.doetempo.models
 
 data class Count(
-    val supportedCampaigns: Long? = null,
-    val following: Long? = null
+    val supportedCampaigns: Int? = null,
+    val following: Int? = null
 )
